moved validation function to new file
